streamlit_app/app.py
import streamlit as st
import numpy as np
import pandas as pd
import joblib
from pathlib import Path
import plotly.graph_objects as go
import plotly.express as px
from PIL import Image
import logging
import yaml
import os
import sys

# nolll's Configuration dynamique du projet
CHEMIN_RACINE = Path(__file__).parent.parent
sys.path.append(str(CHEMIN_RACINE))

# Importations locales du code source nettoyé
from src.preprocessing import PreprocesseurTexte
from src.feature_engineering import PipelineFeatures
from src.modeling import EvaluateurModele

# Configuration du logging
logging.basicConfig(level=logging.INFO)
journal = logging.getLogger(__name__)

# Chargement du fichier de configuration
try:
    with open(CHEMIN_RACINE / "config" / "config.yaml", "r", encoding="utf-8") as f:
        config = yaml.safe_load(f)
except Exception as e:
    st.error(f"Erreur de chargement du fichier de configuration : {e}")
    config = {}

MODELS_DIR = CHEMIN_RACINE / "models" / "trained_models"
DATA_DIR = CHEMIN_RACINE / "data"

def charger_meilleur_modele():
    """Charge le meilleur modèle entraîné."""
    try:
        # Recherche du fichier .joblib ou .pkl
        fichiers = list(MODELS_DIR.glob("*.joblib")) + list(MODELS_DIR.glob("*.pkl"))
        if not fichiers:
            return None
        # On prend le plus récent ou un nom spécifique si défini dans config
        meilleur_modele_chemin = fichiers[0] 
        return joblib.load(meilleur_modele_chemin)
    except Exception as e:
        journal.error(f"Erreur de chargement du modèle : {e}")
        return None

def charger_donnees_entrainement():
    """Charge les données d'entraînement pour les statistiques."""
    try:
        # Chemins basés sur la structure du projet clean
        # Note: Ces fichiers doivent exister après l'exécution du notebook de preprocessing
        chemin_x = DATA_DIR / "csv" / "X_train_processed.npy"
        chemin_y = DATA_DIR / "csv" / "y_train.npy"
        if chemin_x.exists() and chemin_y.exists():
            return np.load(chemin_x), np.load(chemin_y)
        return None, None
    except Exception as e:
        journal.error(f"Erreur de chargement des données : {e}")
        return None, None

# Configuration de la page Streamlit
st.set_page_config(page_title="Classification de Produits Rakuten", layout="wide")
st.title("Classification de Produits Rakuten")

# Barre latérale (Sidebar)
with st.sidebar:
    st.header("Navigation")
    page = st.radio(
        "Choisir une page :",
        ["Tableau de bord EDA", "Prédictions", "Performance du Modèle", "Documentation"]
    )

# # # # # # # # # # # # # # # # # # # # # # # #
# PAGE 1 : TABLEAU DE BORD EDA
# # # # # # # # # # # # # # # # # # # # # # # #
if page == "Tableau de bord EDA":
    st.header("Analyse Exploratoire des Données (EDA)")
    
    X_entrainement, y_entrainement = charger_donnees_entrainement()
    
    if X_entrainement is not None and y_entrainement is not None:
        col1, col2, col3 = st.columns(3)
        
        with col1:
            st.metric("Total des échantillons", len(X_entrainement))
        with col2:
            st.metric("Nombre de caractéristiques", X_entrainement.shape[1])
        with col3:
            st.metric("Nombre de classes", len(np.unique(y_entrainement)))
        
        st.divider()
        
        # Distribution des classes
        unique, counts = np.unique(y_entrainement, return_counts=True)
        df_dist = pd.DataFrame({
            'Classe': unique.astype(str),
            'Nombre': counts,
            'Pourcentage': (counts / len(y_entrainement) * 100).round(2)
        })
        
        fig = px.bar(df_dist, x='Classe', y='Nombre', title="Distribution des Classes", color='Nombre',
                     color_continuous_scale='Viridis')
        st.plotly_chart(fig, use_container_width=True)
        
        st.subheader("Détails des données d'entraînement")
        st.dataframe(df_dist, use_container_width=True)
    else:
        st.info("Les données d'entraînement traitées ne sont pas disponibles. Exécutez le pipeline de prétraitement pour générer les fichiers .npy dans data/csv/.")

# # # # # # # # # # # # # # # # # # # # # # # #
# PAGE 2 : PRÉDICTIONS
# # # # # # # # # # # # # # # # # # # # # # # #
elif page == "Prédictions":
    st.header("Faire une Prédiction")
    
    modele = charger_meilleur_modele()
    
    if modele:
        st.info("Saisissez les informations du produit pour obtenir une classification")
        
        col1, col2 = st.columns(2)
        
        with col1:
            desig_input = st.text_input("Désignation du produit :")
            desc_input = st.text_area("Description du produit :", height=150)
        
        with col2:
            # Placeholder pour l'image si supportée plus tard
            st.write("Image du produit")
            fichier_image = st.file_uploader("Téléchargez une image (optionnel)", type=["jpg", "png", "jpeg"])
            if fichier_image:
                st.image(fichier_image, caption="Image téléchargée", width=200)
        
        if st.button("Lancer la Prédiction", use_container_width=True):
            if desig_input:
                try:
                    # ediz's Note: En production, il faudrait charger le pipeline de features complet (TF-IDF/SVD)
                    # Ici nous simulons la réponse pour la démo de l'UI si les préprocesseurs pkl ne sont pas là
                    st.success("Prédiction effectuée avec succès")
                    
                    # Simulation de résultat (à remplacer par pipeline.transformer si chargé)
                    col_res1, col_res2 = st.columns(2)
                    with col_res1:
                        st.subheader("Classe Prédite")
                        st.code("10 (Livres)")
                    with col_res2:
                        st.subheader("Confiance")
                        st.progress(0.92)
                        st.write("92.4%")
                except Exception as e:
                    st.error(f"Une erreur est survenue lors de la prédiction : {e}")
            else:
                st.warning("Veuillez au moins saisir une désignation de produit.")
    else:
        st.error("Aucun modèle entraîné n'a été trouvé dans models/trained_models/. Veuillez lancer l'entraînement des modèles d'abord.")

# # # # # # # # # # # # # # # # # # # # # # # #
# PAGE 3 : PERFORMANCE
# # # # # # # # # # # # # # # # # # # # # # # #
elif page == "Performance du Modèle":
    st.header("Métriques de Performance")
    
    chemin_eval = CHEMIN_RACINE / "reports" / "evaluation_modeles.csv"
    
    if chemin_eval.exists():
        df_resultats = pd.read_csv(chemin_eval)
        st.subheader("Comparaison des modèles")
        st.dataframe(df_resultats, use_container_width=True)
        
        # Graphique des scores F1
        if "f1_macro" in df_resultats.columns and "modele" in df_resultats.columns:
            fig = px.bar(df_resultats, x='modele', y='f1_macro', title="F1-Score Macro par Modèle",
                         color='f1_macro', color_continuous_scale='Plasma')
            st.plotly_chart(fig, use_container_width=True)
    else:
        st.warning("Aucun rapport d'évaluation trouvé dans reports/. Veuillez exécuter le script ou le notebook d'évaluation finale.")
        st.info("Le fichier attendu est : reports/evaluation_modeles.csv")

# # # # # # # # # # # # # # # # # # # # # # # #
# PAGE 4 : DOCUMENTATION
# # # # # # # # # # # # # # # # # # # # # # # #
else:
    st.header("Documentation du Projet")
    
    st.markdown("""
    ### Objectif du Projet
    Classifier automatiquement les produits du catalogue Rakuten dans 27 catégories différentes en utilisant des données textuelles et des images.

    ### Architecture du Pipeline
    1.  **Prétraitement** : Nettoyage du texte, normalisation, gestion des valeurs manquantes.
    2.  **Ingénierie des Caractéristiques** :
        *   Texte : TF-IDF avec réduction de dimension via SVD (70 à 200 composantes).
        *   Images : Extraction de features via réseaux de neurones pré-entraînés (ResNet/EfficientNet).
        *   Fusion : Combinaison des vecteurs texte et image pour une approche multimodale.
    3.  **Modélisation** :
        *   Baselines : Régression Logistique, Forêt Aléatoire.
        *   Modèles avancés : XGBoost, LightGBM.
        *   Ensemble : Vote majoritaire et pondéré.

    ### Structure des Fichiers
    *   `src/` : Modules Python de base (prétraitement, modélisation).
    *   `notebooks/` : Expérimentations et analyses détaillées.
    *   `scripts/` : Automatisation des tâches répétitives (batch processing).
    *   `data/` : Données brutes et traitées (non présentes sur Git).
    *   `models/` : Modèles sauvegardés et préprocesseurs.
    *   `reports/` : Rapports de performance et visualisations.
    """)
    
    st.divider()
    st.caption("Projet Rakuten - Pipeline de Classification Multimodale | Développé avec soin")

# Pied de page
st.divider()
st.caption(f"Chemin du projet : {CHEMIN_RACINE}")
